requirements.txt
pip == 23.2.1
deta==1.2.0
pandas
pygwalker==0.3.3
python-dotenv==1.0.0
qrcode==7.4.2
reportlab==4.0.4
streamlit==1.25.0
streamlit_authenticator==0.2.2
streamlit_modal==0.1.0
streamlit-aggrid
